src/pyproject.toml
[tool.poetry]
name = "graphiti"
version = "0.1.0"
description = "Zep Graphiti"
authors = ["EM <test@test.com>"]
readme = "README.md"
packages = [{include = "app"}]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = "0.29.*"
fastapi = "0.111.*"
requests = "^2.32.3"
python-dotenv = "^1.0.0"
python-json-logger = "^2.0.7"
prometheus-client = "^0.20.0"
ddtrace = "^2.11.3"
httpx = "^0.27.0"
graphiti-core = "^0.8.2"

[tool.poetry.group.test.dependencies]
pytest-dotenv = "^0.5.2"
pytest = "^8.1.1"
parameterized = "^0.9.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
